Refactor: #32 - Refactor 변경 사항 적용
./api/common/properties, ./api/common/config -> ./domain/common/properties, ./domain/common/config 이동
